% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{add_image_shadow}
\alias{add_image_shadow}
\title{add_image_shadow}
\usage{
add_image_shadow(path_image)
}
\arguments{
\item{path_image}{character, path name of image}
}
\value{
character, path name of image returned invisibly
}
\description{
This function adds a border shadow to the image, archiving the
original image in a separate archive sub-directory
}
\examples{
\dontrun{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
  library(palmerpenguins)
  library(fs)
  library(dplyr)
  library(ggplot2)
  library(here)
  library(magick)
  library(future)
 library(furrr)
 })
})
# create output directory
i_am("example.Rmd")
if (!dir_exists("output")) {dir.create("output")}
if (!dir_exists("output/images")) {dir.create("output/images")}

# example taken from palmerpenguins example analysis of mass vs. flipper length
# https://allisonhorst.github.io/palmerpenguins/articles/examples.html

# data
mass_flipper_data <- penguins \%>\%
  select(species, flipper_length_mm, body_mass_g, )
mass_flipper_data

# graph
mass_flipper_graph <- ggplot(data = mass_flipper_data,
                             aes(x = flipper_length_mm,
                                 y = body_mass_g)) +
  geom_point(aes(color = species,
                 shape = species),
             size = 3,
             alpha = 0.8) +
  theme_minimal() +
  scale_color_manual(values = c("darkorange","purple","cyan4")) +
  labs(title = "Penguin size, Palmer Station LTER",
       subtitle = "Flipper length and body mass for Adelie, Chinstrap and Gentoo Penguins",
       x = "Flipper length (mm)",
       y = "Body mass (g)",
       color = "Penguin species",
       shape = "Penguin species") +
  theme(legend.position = c(0.2, 0.7),
        legend.background = element_rect(fill = "white", color = NA),
        plot.title.position = "plot",
        plot.caption = element_text(hjust = 0, face= "italic"),
        plot.caption.position = "plot")

# save graph
ggsave(filename = "output/images/palmerpenguins_graph.png",
       plot = mass_flipper_graph,
       type = "cairo-png",
       width = 6,
       height = 6,
       units = "in",
      dpi = 72)

# graph image
mass_flipper_graph <- image_read(here("output/images/palmerpenguins_graph.png"))
mass_flipper_graph

# add shadows to graph image
plan(multisession)
here("output/images") \%>\%
  dir_ls(., glob = "*.png") \%>\%
  future_map(add_image_shadow,
             .options = furrr_options(seed = TRUE),
              .progress = TRUE)

# shadowed graph
mass_flipper_graph_shadow <- image_read(here("output/images/palmerpenguins_graph.png"))
mass_flipper_graph_shadow
}
}
\seealso{
The example batch runs the function on a graph from the
\href{https://allisonhorst.github.io/palmerpenguins/articles/examples.html}{palmerpenguins package}
and is an example of parallelization using the \href{https://github.com/HenrikBengtsson/future}{future}
and \href{https://davisvaughan.github.io/furrr/}{furrr} packages

Other image manipulation functions: 
\code{\link{reduce_image_size}()}
}
\concept{image manipulation functions}
