% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{as_gridref}
\alias{as_gridref}
\title{Convert a string to a gridref object}
\usage{
as_gridref(x = character())
}
\arguments{
\item{x}{character, grid reference string}
}
\value{
x, grid reference string with class gridref
}
\description{
Convert character string to gridref object
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# convert to gridref class
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::mutate(grid_reference = as_gridref(grid_reference))
}
}
\seealso{
Constructor function for gridref class as described in \href{https://adv-r.hadley.nz/s3.html}{S3}
chapter of Advanced R second additon by Hadley Wickham
}
