% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{build_all_quarto_articles}
\alias{build_all_quarto_articles}
\title{Build all quarto documents in the vignettes folder into html documents}
\usage{
build_all_quarto_articles(...)
}
\description{
This function builds all the quarto documents in the package's vignettes folder
into a html documents, moving these html documents into the articles sub-folder
of the doc folder so that they can be included in package documentation,
created by the pkgdown package.
}
\details{
use this function after running pkgdown's build_site function to create the
article html documents
}
