% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{build_child_project}
\alias{build_child_project}
\title{Build child project}
\usage{
build_child_project()
}
\description{
This function updates the package ready for backing up as a child project, including
creating an up-to-date renv lock file and updating the package version
}
