% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{build_documentation}
\alias{build_documentation}
\title{Create data project documentation}
\usage{
build_documentation()
}
\description{
This function creates the documentation pages for the data project
}
\details{
Currently home page creation results in an unable to read RDS error and has been
commented out
}
