% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{build_single_quarto_article}
\alias{build_single_quarto_article}
\title{Build a single quarto document into a html document}
\usage{
build_single_quarto_article(...)
}
\description{
This function builds a selected quarto document into a html document, moving
the html file into the articles sub-folder of the doc folder so that it can be
included in package documentation, created by the pkgdown package.
}
