% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{check_parquet_file_compression}
\alias{check_parquet_file_compression}
\title{Check parquet file compression}
\usage{
check_parquet_file_compression(
  .data,
  compression_type = c("snappy", "gzip", "brotli", "zstd", "lz4", "lzo")
)
}
\arguments{
\item{.data}{dataset to be checked for amount of compression using parquet format.}

\item{compression_type}{parquet compression types to check, defaults to all compression
types ("snappy", "gzip", "brotli", "zstd", "lz4", "lzo") unless one or more specific
compression types are supplied as a vector.}
}
\value{
message giving each file size and compression percentage
}
\description{
This function checks the amount of compression attained when exporting a dataset using
the Apache parquet file format, by the \href{https://arrow.apache.org/docs/r/}{arrow package}.
It compares parquet file sizes for a dataset, exported uncompressed and compressed using different available
compression types, with the equivalent dataset's file size exported as a csv file.
}
\examples{
data(penguins, package = "palmerpenguins")
check_parquet_file_compression(penguins)
}
