% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{click}
\alias{click}
\title{Increment the tally counter}
\usage{
click()
}
\description{
This function adds or subtracts one from the counter depending on the tally counter
type used, defined within the tally_counter function. It returns the number padded
to a minimum of four characters. Once counter has finished the environment and counter
object are removed.
}
