% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.R
\name{coefplot}
\alias{coefplot}
\title{Plotting for state-trait ordinal regression coefficients}
\usage{
coefplot(
  obj,
  col = NULL,
  bubble = FALSE,
  xlab = "State",
  ylab = "Person-specific median",
  title = NULL,
  range = c(0, 25),
  hjust = 0.5,
  lineheight = 1.2,
  ...
)
}
\arguments{
\item{obj}{A model object of class \code{\link{store}}.}

\item{col}{The colors for the range of the fitted values.}

\item{bubble}{Logical: a bubble plot if \code{TRUE}, otherwise an \code{\link{image}} plot.}

\item{xlab, ylab}{x- and y-axis labels}

\item{title}{Plot title}

\item{range}{Numeric vector of length 2 specifying the minimum and maximum size of the bubbles, if \code{bubble==TRUE}.}

\item{hjust}{Horizontal justification (in [0,1])}

\item{lineheight}{Line height}

\item{\dots}{Other \code{\link{graphical parameters}}.}
}
\value{
None; a plot is created.
}
\description{
Plots coefficients for the various trait-state combinations of a \code{\link{store}} fit.
}
\seealso{
\code{\link{store}}
}
\author{
Prince P. Osei and Philip T. Reiss
}
