% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assurance.R
\name{compare_datasets}
\alias{compare_datasets}
\title{Compare datasets}
\usage{
compare_datasets(old_dataset, new_dataset)
}
\arguments{
\item{old_dataset}{data frame, earlier dataset for comparison}

\item{new_dataset}{data frame, later dataset for comparison}
}
\value{
data frame, compared dataset with new, removed or updated rows highlighted with #
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function compares two versions of a dataset returning the dataset with the added, removed
or changed rows identified, using the \href{https://github.com/edwindj/daff}{daff package}.
The compared dataset can then be exported into an Excel spreadsheet to quickly
identify where values have been changed using conditional formatting, on text
containing #

An initial check is required before using the function to make sure the column
names between the two datasets are identical. This can be done quickly using the
compare function in the \href{https://www.tidyverse.org/blog/2020/10/waldo/}{waldo package}.
}
