% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{count_pdf_document_pages}
\alias{count_pdf_document_pages}
\title{Count pdf document pages}
\usage{
count_pdf_document_pages(pdf_path)
}
\arguments{
\item{pdf_path}{path character string of pdf document}
}
\value{
integer of number of pages in pdf document
}
\description{
This function counts the number of pages in the pdf document
}
\seealso{
Other pdf: 
\code{\link{merge_pdf_documents}()},
\code{\link{remove_pdf_document_pages}()}
}
\concept{pdf}
