% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{counter}
\alias{counter}
\title{Parent R6 class representing a counter
start The starting count number.
end The end count number.
count The current count number.
increment Whether to increase (add) or decrease (subtract) count number.
pad The number of digits for count number.}
\description{
Parent R6 class representing a counter
start The starting count number.
end The end count number.
count The current count number.
increment Whether to increase (add) or decrease (subtract) count number.
pad The number of digits for count number.

Parent R6 class representing a counter
start The starting count number.
end The end count number.
count The current count number.
increment Whether to increase (add) or decrease (subtract) count number.
pad The number of digits for count number.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{status}}{Whether the counter is active (ON) or inactive (OFF).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-counter}{\code{counter$counter()}}
\item \href{#method-display}{\code{counter$display()}}
\item \href{#method-finalize}{\code{counter$finalize()}}
\item \href{#method-clone}{\code{counter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-counter"></a>}}
\if{latex}{\out{\hypertarget{method-counter}{}}}
\subsection{Method \code{counter()}}{
Increments count number depending on increment value (add or subtract). Changes
status to OFF if count number has reached the end count number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{counter$counter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-display"></a>}}
\if{latex}{\out{\hypertarget{method-display}{}}}
\subsection{Method \code{display()}}{
Displays count number padded with zeros.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{counter$display()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Empty function not used on finalizing counter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{counter$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{counter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
