% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{counter_add}
\alias{counter_add}
\title{Child R6 class representing incrementing counter by adding count number.}
\description{
Child R6 class representing incrementing counter by adding count number.

Child R6 class representing incrementing counter by adding count number.
}
\section{Super class}{
\code{\link[store:counter]{store::counter}} -> \code{counter_add}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{counter_add$new()}}
\item \href{#method-clone}{\code{counter_add$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="store" data-topic="counter" data-id="counter">}\href{../../store/html/counter.html#method-counter}{\code{store::counter$counter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="store" data-topic="counter" data-id="display">}\href{../../store/html/counter.html#method-display}{\code{store::counter$display()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="store" data-topic="counter" data-id="finalize">}\href{../../store/html/counter.html#method-finalize}{\code{store::counter$finalize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Increments the counter by adding count number, defining maximum count.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{counter_add$new(limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{Maximum count as an integer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Create a new counter object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{counter_add$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
