% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{create_bounding_box}
\alias{create_bounding_box}
\title{Create bounding box polygon}
\usage{
create_bounding_box(xmin, ymin, xmax, ymax, epsg_number)
}
\arguments{
\item{xmin}{The left corners easting coordinate as an integer or double.}

\item{ymin}{The lower corners northing coordinate as an integer or double.}

\item{xmax}{The right corners easting coordinate as an integer or double.}

\item{ymax}{The upper corners northing coordinate as an integer or double.}

\item{epsg_number}{EPSG number as an integer.}
}
\value{
The bounding box polygon returned as a sf spatial object.
}
\description{
This function creates a bounding box polygon from it's corner coordinates,
setting the coordinate reference system from the supplied \href{https://epsg.org/home.html}{EPSG value}.
}
\examples{
suppressPackageStartupMessages({
  library(store)
})
create_bounding_box(xmin = -10.81,
                    ymin = 49.85,
                    xmax = 2.07,
                    ymax = 60.96,
                    epsg = 4326)
}
\seealso{
The function was created following the approach taken in the stack overflow
\href{https://stackoverflow.com/questions/44335246/polygons-from-coordinates}{polygons from coordinates}.
}
