% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_locked_keyring}
\alias{create_locked_keyring}
\title{create_locked_keyring}
\usage{
create_locked_keyring(keyring_name)
}
\arguments{
\item{keyring_name}{}
}
\value{
character, keyring name
}
\description{
Add a password protected keyring to the Windows Credential Store.
This keyring will be used to store database connection strings.
}
\details{
On creating the keyring you will be asked for a password.
Remember this password as it will be needed everytime you access
the database connection strings stored in this keyring.
}
\examples{
\dontrun{
create_locked_keyring(readline("keyring name: "))
}
}
