% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{delete_key}
\alias{delete_key}
\title{Delete a key from a keyring}
\usage{
delete_key(keyring_name, service_name)
}
\arguments{
\item{keyring_name}{character, keyring name}

\item{service_name}{character, key name}
}
\value{

}
\description{
Remove a key from the keyring by supplying both the keyring name
and name of key to remove
}
\examples{
\dontrun{
delete_key(readline("keyring name: "), readline("service name: "))
}
}
\seealso{
Other keyring functions: 
\code{\link{add_key}()},
\code{\link{create_locked_keyring}()},
\code{\link{get_keyring_status}()},
\code{\link{list_keys}()},
\code{\link{lock_keyring}()},
\code{\link{unlock_keyring}()}
}
\concept{keyring functions}
