% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{delete_keyring}
\alias{delete_keyring}
\title{Delete a keyring}
\usage{
delete_keyring(keyring_name)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring.}
}
\description{
Deletes a keyring from the Windows Credential Store.
}
\seealso{
Other keyring functions: 
\code{\link{add_key}()},
\code{\link{create_locked_keyring}()},
\code{\link{delete_key}()},
\code{\link{get_keyring_status}()},
\code{\link{list_keys}()},
\code{\link{lock_keyring}()},
\code{\link{unlock_keyring}()}
}
\concept{keyring functions}
