% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_table}
\alias{display_table}
\title{display_table}
\usage{
display_table(data, rows = nrow(data))
}
\arguments{
\item{data}{data frame, data to display}

\item{rows}{integer, number of rows to display with total rows as default}
}
\value{
html table to display
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function displays a data frame as a html table using the \href{https://glin.github.io/reactable/}{reactable}
package. The top ten rows in the table are displayed with the options to click
through or display all the rows. The ability to search for table content and
sort columns is also available.
}
\examples{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
   library(palmerpenguins)
   library(dplyr)
 })
})
# display top 5 heaviest penguins from each species on each island
heaviest_penguins <- penguins \%>\%
  select(species, island, body_mass_g) \%>\%
  group_by(species, island) \%>\%
  arrange(desc(body_mass_g)) \%>\%
  slice_head(n = 5) \%>\%
  ungroup()
display_table(heaviest_penguins)
}
\concept{tabular display}
