% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{display_variable_collection}
\alias{display_variable_collection}
\title{Run exploratory data analyses}
\usage{
display_variable_collection(
  .dataset,
  summary = TRUE,
  outliers = TRUE,
  distribution = TRUE,
  correlation = TRUE
)
}
\arguments{
\item{.dataset}{data frame, dataset for exploratory data analysis}

\item{summary}{boolean, display variable summary (FALSE to turn off)}

\item{outliers}{boolean, display outliers (FALSE to turn off)}

\item{distribution}{boolean, display variable distribution (FALSE to turn off)}

\item{correlation}{boolean, display variable correlation (FALSE to turn off)}
}
\value{
dataset returned invisibly
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Export the collection of exploratory data analyses plots and tables. This includes
the summary, distribution and correlation of variables and presence of outliers
exported by default, any of which may be excluded in the export
}
\examples{
\dontrun{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
  library(palmerpenguins)
  library(here)
  library(fs)
 })
})
# create output directory
i_am("example.Rmd")
if (!dir_exists("output")) {dir.create("output")}

# example from palmerpenguins
# https://allisonhorst.github.io/palmerpenguins/reference/penguins_raw.html
display_variable_collection(penguins_raw)
}
}
\seealso{
This is an example of exploratory data analsyis using the \href{https://github.com/choonghyunryu/dlookr}{dlookr}
and \href{https://alastairrushworth.github.io/inspectdf/}{inspectdf} packages

Other exploratory data analysis: 
\code{\link{display_variable_correlation}()},
\code{\link{display_variable_distribution}()},
\code{\link{display_variable_outliers}()},
\code{\link{display_variable_summary}()}
}
\concept{exploratory data analysis}
