% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{display_variable_distribution}
\alias{display_variable_distribution}
\title{display_variable_distribution}
\usage{
display_variable_distribution(.dataset, height_1 = 1, height_2 = 0)
}
\arguments{
\item{.dataset}{data frame, data frame used to display the variable statistics}

\item{height_1}{integer, integer between 0-1, determining height of first column}

\item{height_2}{integer, integer between 0-1, determining height of second column}
}
\value{
a patchwork object including one plot and one table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Displays the distribution of the numeric variables within a data frame, including
a histogram for each variable and table of summary statistics. These statistics
include the range, quartiles, mean, medium, standard deviation, standard error of
the mean, level of skewness, kurtosis and normality
}
\seealso{
Other exploratory data analysis: 
\code{\link{display_variable_correlation}()},
\code{\link{display_variable_outliers}()},
\code{\link{display_variable_summary}()}
}
\concept{exploratory data analysis}
