% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{display_variable_outliers}
\alias{display_variable_outliers}
\title{display_variable_outliers}
\usage{
display_variable_outliers(.dataset, width_1 = 0.6, width_2 = 0.4)
}
\arguments{
\item{.dataset}{data frame, data frame used to display the variable statistics}

\item{width_1}{integer, integer between 0-1, determining width of first column}

\item{width_2}{integer, integer between 0-1, determining width of second column}
}
\value{
a patchwork object including one plot and one table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Displays the variable outliers within a data frame with a boxplot, and
summary statistics including count outliers and mean of each variable
with outliers included and excluded
}
\seealso{
Other exploratory data analysis: 
\code{\link{display_variable_correlation}()},
\code{\link{display_variable_distribution}()},
\code{\link{display_variable_stats}()}
}
\concept{exploratory data analysis}
