% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{display_variable_stats}
\alias{display_variable_stats}
\title{display_variable_stats}
\usage{
display_variable_stats(.dataset, width_1 = 0.6, width_2 = 0.4)
}
\arguments{
\item{.dataset}{data frame, data frame used to display the variable statistics}

\item{width_1}{integer, integer between 0-1, determining width of first column}

\item{width_2}{integer, integer between 0-1, determining width of second column}
}
\value{
a patchwork object including two plots and one table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Displays the basic statistics of the variables within a data frame. This includes
a plot of the prevalence of NAs and frequency of category levels, a summary count
of the number of rows, missing values, unique values and zero values for each variable
}
\seealso{
Other exploratory data analysis: 
\code{\link{display_variable_correlation}()},
\code{\link{display_variable_distribution}()},
\code{\link{display_variable_outliers}()}
}
\concept{exploratory data analysis}
