% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{display_variable_summary}
\alias{display_variable_summary}
\title{display_variable_summary}
\usage{
display_variable_summary(.dataset)
}
\arguments{
\item{.dataset}{data frame, data frame to summarise}
}
\value{
dataset returned invisibly
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Exports the basic statistics of the variables within a data frame into an output
directory. This includes a plots of the prevalence of missing values and frequency
of category levels and a table image containing variable's type summary count of
the number of rows, missing values, unique values and zero values for each variable
}
\examples{
\dontrun{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
  library(palmerpenguins)
  library(here)
  library(fs)
 })
})
# create output directory
i_am("example.Rmd")
if (!dir_exists("output")) {dir.create("output")}

# example from palmerpenguins
# https://allisonhorst.github.io/palmerpenguins/reference/penguins_raw.html
display_variable_summary(penguins_raw)
}
}
\seealso{
Other exploratory data analysis: 
\code{\link{display_variable_correlation}()},
\code{\link{display_variable_distribution}()},
\code{\link{display_variable_outliers}()}
}
\concept{exploratory data analysis}
