% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{easting.gridref}
\alias{easting.gridref}
\title{Get easting for grid reference}
\usage{
\method{easting}{gridref}(grid_reference, centre = FALSE)
}
\arguments{
\item{grid_reference}{character gridref class, British or Irish grid reference}

\item{centre}{logical, easting for either lower left hand corner (default) or centre point (TRUE)}
}
\value{
integer, easting of grid reference in metres.
}
\description{
This function returns the grid reference's easting in metres.
It uses the gridCoords function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add easting column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::mutate(grid_reference = as_gridref(grid_reference)) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(easting = easting(grid_reference, centre = FALSE))
}
}
\seealso{
Other grid reference functions: 
\code{\link{gridsquare_geometry.gridref}()},
\code{\link{hectad.gridref}()},
\code{\link{hectare.gridref}()},
\code{\link{monad.gridref}()},
\code{\link{northing.gridref}()},
\code{\link{pentad.gridref}()},
\code{\link{precision.gridref}()},
\code{\link{projection.gridref}()},
\code{\link{tetrad.gridref}()}
}
\concept{grid reference functions}
