% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{easting}
\alias{easting}
\title{Generic function for easting}
\usage{
easting(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by generic function easting}
}
\value{

}
\description{
Generic function for easting as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
