% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{eda}
\alias{eda}
\title{Run the exploratory data analyses}
\usage{
eda(
  .dataset,
  summary = TRUE,
  outliers = TRUE,
  distribution = TRUE,
  correlation = TRUE
)
}
\arguments{
\item{.dataset}{A data frame requiring exploratory data analysis}

\item{summary}{Use the default TRUE to display variable summary or FALSE to not display variable summary}

\item{outliers}{Use the default TRUE to display outliers or FALSE to not display outliers}

\item{distribution}{Use the default TRUE to display variable distribution or FALSE to not display variable distribution}

\item{correlation}{Use the default TRUE to display variable correlation orFALSE to not display variable correlation}
}
\value{
The data frame is returned invisibly so that the function can be
used in a piped workflow
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Export the collection of exploratory data analyses plots and tables. This includes
the summary, distribution and correlation of variables and presence of outliers
exported by default, any of which may be excluded in the export.
}
\details{
The output directory is created in the project root directory which may be changed
using here::i_am function.
}
\examples{
\dontrun{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
  library(palmerpenguins)
  library(here)
  library(fs)
 })
})
# create output directory
i_am("example.Rmd")
if (!dir_exists("output")) {dir_create("output")}

# example from palmerpenguins
# https://allisonhorst.github.io/palmerpenguins/reference/penguins_raw.html
eda_variable_collection(penguins_raw)
}
}
\seealso{
Other exploratory data analysis: 
\code{\link{eda_variable_correlation}()},
\code{\link{eda_variable_distribution}()},
\code{\link{eda_variable_outliers}()},
\code{\link{eda_variable_summary}()}
}
\concept{exploratory data analysis}
