% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{eda_variable_distribution}
\alias{eda_variable_distribution}
\title{Display variable distribution}
\usage{
eda_variable_distribution(.dataset)
}
\arguments{
\item{.dataset}{data frame, dataset for exploratory data analysis}
}
\value{
dataset returned invisibly
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Exports the distribution of the numeric variables within a data frame, including
a histogram for each variable and table of summary statistics. These statistics
include the range, quartiles, mean, medium, standard deviation, standard error of
the mean, level of skewness, kurtosis and normality

The output directory is created in the project root directory which may be changed
using here::i_am function
}
\examples{
\dontrun{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
  library(palmerpenguins)
  library(here)
  library(fs)
 })
})
# create output directory
i_am("example.Rmd")
if (!dir_exists("output")) {dir_create("output")}

# example from palmerpenguins
# https://allisonhorst.github.io/palmerpenguins/reference/penguins_raw.html
eda_variable_distribution(penguins_raw)
}
}
\seealso{
This is an example of exploratory data analsyis using the \href{https://github.com/choonghyunryu/dlookr}{dlookr}
and \href{https://alastairrushworth.github.io/inspectdf/}{inspectdf} packages

Other exploratory data analysis: 
\code{\link{eda_variable_correlation}()},
\code{\link{eda_variable_outliers}()},
\code{\link{eda_variable_summary}()},
\code{\link{eda}()}
}
\concept{exploratory data analysis}
