% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{eda_variable_summary}
\alias{eda_variable_summary}
\title{Display a summary of the variables}
\usage{
eda_variable_summary(.dataset)
}
\arguments{
\item{.dataset}{data frame, dataset for exploratory data analysis}
}
\value{
dataset returned invisibly
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Exports the basic statistics of the variables within a data frame into an output
directory. This includes a plots of the prevalence of missing values and frequency
of category levels and a table image containing variable's type summary count of
the number of rows, missing values, unique values and zero values for each variable

The output directory is created in the project root directory which may be changed
using here::i_am function
}
\examples{
\dontrun{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
  library(palmerpenguins)
  library(here)
  library(fs)
 })
})
# create output directory
i_am("example.Rmd")
if (!dir_exists("output")) {dir_create("output")}

# example from palmerpenguins
# https://allisonhorst.github.io/palmerpenguins/reference/penguins_raw.html
eda_variable_summary(penguins_raw)
}
}
\seealso{
This is an example of exploratory data analsyis using the \href{https://github.com/choonghyunryu/dlookr}{dlookr}
and \href{https://alastairrushworth.github.io/inspectdf/}{inspectdf} packages

Other exploratory data analysis: 
\code{\link{eda_variable_correlation}()},
\code{\link{eda_variable_distribution}()},
\code{\link{eda_variable_outliers}()},
\code{\link{eda}()}
}
\concept{exploratory data analysis}
