% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_spatial_data.R
\docType{data}
\name{example_lines}
\alias{example_lines}
\title{example_lines}
\format{
A spatial object with 5 attribute rows and 3 columns.
\describe{
\item{id}{integer id of linestring.}
\item{relation}{spatial relationship of linestrings to the example main polygon.}
}
}
\usage{
example_lines
}
\description{
Five parallel lines in British National Grid projection. These lines either
overlaps, touches or are outside the example_main_polygon.
}
\section{Formats}{

The spatial object can be saved in parquet and shape file format by running the code in
the example_data. R script stored in the data_raw folder. The files are saved
in the inst/extdata directory.
}

\keyword{datasets}
