% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_spatial_data.R
\docType{data}
\name{example_main_polygon}
\alias{example_main_polygon}
\title{example_main_polygon}
\format{
A spatial object with 1 attribute row and 2 columns.
\describe{
\item{id}{integer id of polygon.}
}
}
\usage{
example_main_polygon
}
\description{
One KM square polygon in British National Grid projection.
}
\section{Formats}{

The spatial object can be saved in parquet and shape file format by running the code in
the example_data. R script stored in the data_raw folder. The files are saved
in the inst/extdata directory.
}

\keyword{datasets}
