% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_spatial_data.R
\docType{data}
\name{example_multipoint}
\alias{example_multipoint}
\title{example_multipoint}
\format{
A spatial object with 1 attribute rows and 1 columns.
\describe{
\item{FID}{integer id of multipoint.}
}
}
\usage{
example_multipoint
}
\description{
One multipoint created from the twenty example points in British National Grid
projection.
}
\section{Formats}{

The spatial object can be saved in parquet and shape file format by running the code in
the example_data. R script stored in the data_raw folder. The files are saved
in the inst/extdata directory.
}

\keyword{datasets}
