% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_spatial_data.R
\docType{data}
\name{example_multipolygon}
\alias{example_multipolygon}
\title{example_multipolygon}
\format{
A spatial object with 1 attribute rows and 1 columns.
\describe{
\item{FID}{integer id of multipolygon.}
}
}
\usage{
example_multipolygon
}
\description{
One multipolygon created from the four example polygons in British National Grid
projection, using st_union to resolve the polygon boundaries.
}
\section{Formats}{

The spatial object can be saved in parquet and shape file format by running the code in
the example_data. R script stored in the data_raw folder. The files are saved
in the inst/extdata directory.
}

\keyword{datasets}
