% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_spatial_data.R
\docType{data}
\name{example_points}
\alias{example_points}
\title{example_points}
\format{
A spatial object with 20 attribute rows and 3 columns.
\describe{
\item{id}{integer id of multipolygon.}
\item{relation}{spatial relationship of points to the example main polygon.}
}
}
\usage{
example_points
}
\description{
Twenty points, consisting of five rows of four, in British National Grid projection.
These points either are contain within, touches or are outside the example_main_polygon.
}
\section{Formats}{

The spatial object can be saved in parquet and shape file format by running the code in
the example_data. R script stored in the data_raw folder. The files are saved
in the inst/extdata directory.
}

\keyword{datasets}
