% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_spatial_data.R
\docType{data}
\name{example_polygons}
\alias{example_polygons}
\title{example_polygons}
\format{
A spatial object with 4 attribute rows and 4 columns.
\describe{
\item{id}{integer id of polygon.}
\item{relation}{spatial relationship of polygon to the example main polygon.}
\item{DE-9IM}{The Dimensionally Extended 9-Intersection Model used to describe the
spatial relationship between the example polygon and the example main polygon.}
}
}
\usage{
example_polygons
}
\description{
Four 500m square polygons in British National Grid projection. These polygons
either overlap, are contain within, touches or are outside the example_main_polygon.
}
\section{Formats}{

The spatial object can be saved in parquet and shape file format by running the code in
the example_data. R script stored in the data_raw folder. The files are saved
in the inst/extdata directory.
}

\keyword{datasets}
