% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{extract_delimited_list}
\alias{extract_delimited_list}
\title{extract_delimited_list}
\usage{
extract_delimited_list(delimiter)
}
\arguments{
\item{delimiter}{character, delimiter}
}
\value{
function, extracts the column as a delimited list
}
\description{
This function returns the parent function of the function factory to extract
a delimited list from a data frame's column. It excludes NAs but does not sort
the character string
}
