% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{extract_paragraph_delimited_list}
\alias{extract_paragraph_delimited_list}
\title{Extract a paragraph delimited list of items}
\usage{
extract_paragraph_delimited_list(
  .df,
  column_name,
  sort = FALSE,
  unique_list = FALSE,
  last_delimiter = "",
  ...
)
}
\arguments{
\item{.df}{A data frame containing the column with the delimited list to extract.}

\item{column_name}{The name of the column containing the delimited list to extract.}

\item{sort}{Should the delimited list be sorted (TRUE) or not (default FALSE).}

\item{unique_list}{Should the delimited list contain duplicated elements (default FALSE) or not (TRUE).}

\item{last_delimiter}{An optional character string used to separate the last two items in the delimited list.}
}
\value{
The paragraph separated delimited list as a character string.
}
\description{
This function extracts a paragraph delimited list from a data frame's column. It
excludes NAs, optionally removes duplicate elements, sorts the character string
and adds text as a delimiter to the last two items.
}
\details{
It is an example of the use of a \href{https://adv-r.hadley.nz/function-factories.html}{function factory},
\href{https://www.tidyverse.org/blog/2020/02/glue-strings-and-tidy-eval/}{tidy evaluation}
and purrr's \href{https://purrr.tidyverse.org/reference/map.html}{map function}.
The function may be called on a \href{https://cran.r-project.org/web/packages/tidyr/vignettes/nest.html}{nested data frame}
to extract the delimited list.
}
\examples{
suppressPackageStartupMessages({
  library(store)
  suppressWarnings({
    library(palmerpenguins)
    library(dplyr)
  })
})

# select top 5 heaviest penguins from each species on each island
heaviest_penguins <- penguins \%>\%
  select(species, island, body_mass_g) \%>\%
  group_by(species, island) \%>\%
  arrange(desc(body_mass_g)) \%>\%
  slice_head(n = 5) \%>\%
  ungroup()
heaviest_penguins

# extract paragraph separated list of penguin weights for each species on each island
suppressPackageStartupMessages({
  suppressWarnings({
   library(purrr)
  })
})

heaviest_penguins \%>\%
 group_nest(across(c(species:island)), .key = "penguins") \%>\%
 mutate(weight = map_chr(penguins, extract_paragraph_delimited_list, column_name = "body_mass_g")) \%>\%
 select(-penguins)
}
\seealso{
Other delimited functions: 
\code{\link{extract_comma_delimited_list}()},
\code{\link{extract_semicolon_delimited_list}()},
\code{\link{extract_space_delimited_list}()}
}
\concept{delimited functions}
