% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{extract_polygons}
\alias{extract_polygons}
\title{Extract and dissolve polygons from geometry collection}
\usage{
extract_polygons(sf_data)
}
\arguments{
\item{sf_data}{sf object, spatial data frame containing geometry collections}
}
\value{
sf object, spatial data frame containing extracted multipolygons
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function extracts polygons from a geometry collection, dissolving them into
multipolygons
}
\examples{
\dontrun{
extract_polygons(uk_ireland_base_map)
}
}
