% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{extract_semicolon_delimited_list}
\alias{extract_semicolon_delimited_list}
\title{Extract a semi-colon delimited list of items

This function extracts a semi-colon list from a data frame's column. It
excludes NAs, optionally sorts the character string and adds text as a delimiter
to the last two items}
\usage{
extract_semicolon_delimited_list(
  .df,
  column_name,
  sort = FALSE,
  unique_list = FALSE,
  ...
)
}
\arguments{
\item{.df}{data frame, data frame containing the column to extract the delimited list}

\item{column_name}{character, name of column to extract the delimited list}

\item{sort}{logical, whether the list is sorted (TRUE) or not (default FALSE)}

\item{unique_list}{logical, whether to remove the duplicated elements (TRUE) or not (default FALSE)}

\item{last, }{character, optional string used to separate the last two items}
}
\value{
character, semi-colon delimited list
}
\description{
It is an example of the use of a \href{https://adv-r.hadley.nz/function-factories.html}{function factory},
\href{https://www.tidyverse.org/blog/2020/02/glue-strings-and-tidy-eval/}{tidy evaluation}
and purrr's \href{https://purrr.tidyverse.org/reference/map.html}{map function}.
The function may be called on a \href{https://cran.r-project.org/web/packages/tidyr/vignettes/nest.html}{nested data frame}
to extract the semi-colon delimited list
}
\examples{
suppressPackageStartupMessages({
  library(store)
  suppressWarnings({
    library(palmerpenguins)
    library(dplyr)
    library(purrr)
  })
})
# select top 5 heaviest penguins from each species on each island
heaviest_penguins <- penguins \%>\%
 select(species, island, body_mass_g) \%>\%
 group_by(species, island) \%>\%
 arrange(desc(body_mass_g)) \%>\%
 slice_head(n = 5) \%>\%
 ungroup()
heaviest_penguins

# extract comma separated list of penguin weights for each species on each island
heaviest_penguins <- heaviest_penguins \%>\%
 group_nest(across(c(species:island)), .key = "penguins") \%>\%
 mutate(weight = map_chr(penguins, extract_semicolon_delimited_list, column_name = "body_mass_g")) \%>\%
 select(-penguins)
heaviest_penguins
}
\seealso{
Other delimited functions: 
\code{\link{extract_comma_delimited_list}()},
\code{\link{extract_paragraph_delimited_list}()},
\code{\link{extract_space_delimited_list}()}
}
\concept{delimited functions}
