% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_system.R
\name{filenames_to_clipboard}
\alias{filenames_to_clipboard}
\title{Copy list of file names to clipboard}
\usage{
filenames_to_clipboard()
}
\value{
console message reporting list of files names copied to the clipboard
}
\description{
This function copies the list of file names contained within a directory and
it's child directories to the clipboard.
}
