% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitplot.R
\name{fitplot}
\alias{fitplot}
\title{Plotting for state-trait ordinal regression fitted values}
\usage{
fitplot(
  obj,
  col = NULL,
  bubble = FALSE,
  xlab = "State",
  ylab = "Person-specific median",
  title = NULL,
  range = c(0, 25),
  hjust = 0.5,
  lineheight = 1.2,
  ...
)
}
\arguments{
\item{obj}{A model object of class \code{\link{store}}.}

\item{col}{The colors for the range of the fitted values.}

\item{bubble}{Logical: a bubble plot if \code{TRUE}, otherwise an \code{\link{image}} plot.}

\item{xlab, ylab}{x- and y-axis labels}

\item{title}{Plot title}

\item{range}{Scaling values (in [0,25])}

\item{hjust}{Horizontal justification (in [0,1])}

\item{lineheight}{Line height}

\item{\dots}{Other \code{\link{graphical parameters}}.}
}
\value{
None; a plot is created.
}
\description{
Plots fitted values for the various state-trait combinations for a \code{\link{store}} fit.
}
\seealso{
\code{\link{store}}
}
\author{
Prince P. Osei and Philip T. Reiss
}
