% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{get_cran_package_system_dependencies}
\alias{get_cran_package_system_dependencies}
\title{Get CRAN packages' system dependencies}
\usage{
get_cran_package_system_dependencies(
  renv_lock_path,
  distribution = "ubuntu",
  release = NULL
)
}
\arguments{
\item{renv_lock_path}{Path to the renv.lock file.}

\item{distribution}{Operating system to be used by the Docker image, with ubuntu as
the default. Other operating systems include centos.}

\item{release}{Release version of the operating system. This may default to the latest
operating system release.}
}
\value{
An httr response, containing the packages requiring system dependencies, the
system dependencies required and their install command, within a list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function extracts the list of CRAN packages within a renv lock file, querying
the \href{https://packagemanager.rstudio.com/__api__/swagger/index.html}{RStudio Package Manager API}
to get a list of external system libraries these packages depend on.

The returned list of system dependencies are intended to help create a Docker file
used to build a Docker image for a R project.
}
\examples{
\dontrun{
get_cran_package_system_dependencies(renv_lock_path = "renv.lock",
                                     distribution = "unbuntu",
                                     release = "20.04")
}
}
\seealso{
The function is a collation and modification of the functions in the \href{https://github.com/mdneuzerling/getsysreqs/tree/main/R}{getsysres package}
written by David Neuzerling.
}
