% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_keyring_status}
\alias{get_keyring_status}
\title{Get the status of a keyring}
\usage{
get_keyring_status(keyring_name)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring}
}
\value{
A message giving the locked status of the keyring
}
\description{
Check the current status of the keyring. The default status of the keyring is to be locked,
requiring its password to unlock it, to access the stored database connection strings.
}
\seealso{
Other keyring functions: 
\code{\link{add_key}()},
\code{\link{create_locked_keyring}()},
\code{\link{delete_key}()},
\code{\link{list_keys}()},
\code{\link{lock_keyring}()},
\code{\link{unlock_keyring}()}
}
\concept{keyring functions}
