% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_multipolygon_spatial_dimensions}
\alias{get_multipolygon_spatial_dimensions}
\title{Get multipolygon spatial dimensions}
\usage{
get_multipolygon_spatial_dimensions(.data)
}
\arguments{
\item{.data}{spatial data frame containing a feature_code and feature_name columns,
along with a sf geometry column and co-ordinate reference system}
}
\value{
data frame containing original feature_code and feature_name fields, along with
calculated EPSG code, area, length and x and y centroid co-ordinates.
}
\description{
This functions returns the co-ordinate reference system, area, perimeter length
and x and y centroid co-ordinates of a spatial data frame containing multipolygons.
It may be used to compare difference versions of a spatial dataset.
}
\examples{
\dontrun{
get_multipolygon_spatial_dimensions(sf_data_frame)
}
}
