% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_mysql_connection}
\alias{get_mysql_connection}
\title{get_mysql_connection}
\usage{
get_mysql_connection(keyring_name, service_name)
}
\arguments{
\item{keyring_name}{character, keyring name}

\item{service_name}{character, name of service storing credentials}
}
\value{
connection object, MySQL connection object
}
\description{
This function initiates a MySQL connection using stored database credentials
using the keyring package. The MySQL tables are added to the connection pane
}
\examples{
\dontrun{
get_mysql_connection("keyring", "database_service_name")
}
}
