% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_postgres_connection}
\alias{get_postgres_connection}
\title{get_postgres_connection}
\usage{
get_postgres_connection(keyring_name, service_name)
}
\arguments{
\item{keyring_name}{character, keyring name}

\item{service_name}{name of service storing credentials}
}
\value{
connection object, PostgreSQL connection object
}
\description{
This function initiates a PostgreSQL connection using stored database credentials
through the keyring package. The PostgreSQL tables are added to the connection pane.
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
  suppressWarnings({
    library(DBI)
    library(odbc)
  })
})
# connect to database, using keyring stored credentials
con <- get_postgres_connection(keyring_name = "keyring_name",
                               service_name = "database_name")
# list tables
dbListTables(con)

# get query
database_query <- dbGetQuery(con, "sql_statement")

# import table
dbWriteTable(con, Id(schema = "schema_name", table = "table_name"),
             value = data_frame, append = TRUE, row.names = FALSE)

# disconnect from database
dbDisconnect(con)
}
}
\seealso{
More information on connecting to databases using R can be found in RStudio's
\href{https://db.rstudio.com/}{Databases using R} webpages.

Other database functions: 
\code{\link{get_mysql_connection}()},
\code{\link{get_sqlserver_connection}()}
}
\concept{database functions}
