% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{grid_reference_easting}
\alias{grid_reference_easting}
\title{Get easting for grid reference}
\usage{
grid_reference_easting(grid_reference, centre = FALSE)
}
\arguments{
\item{grid_reference}{character, British or Irish grid reference}

\item{centre}{logical, easting for either lower left hand corner (default) or centre point (TRUE)}
}
\value{
integer, easting of grid reference in metres.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function returns the grid reference's easting in metres.
It uses the gridCoords function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.

It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add easting column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(easting = grid_reference_easting(grid_reference, centre = FALSE))
}
}
\seealso{
Other grid reference functions: 
\code{\link{grid_reference_geometry}()},
\code{\link{grid_reference_hectad}()},
\code{\link{grid_reference_hectare}()},
\code{\link{grid_reference_monad}()},
\code{\link{grid_reference_northing}()},
\code{\link{grid_reference_pentad}()},
\code{\link{grid_reference_projection}()},
\code{\link{grid_reference_tetrad}()},
\code{\link{precision.gridref}()}
}
\concept{grid reference functions}
