% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{grid_reference_geometry}
\alias{grid_reference_geometry}
\title{Convert OSGB or OSNI Grid reference to polygon geometry feature}
\usage{
grid_reference_geometry(grid_reference)
}
\arguments{
\item{grid_reference}{character, British or Irish grid reference}
}
\value{
geometry, square polygon feature
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function converts a grid reference to its square polygon geometry feature
through conversion to well-known text. It uses the gridCoords function in the
archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.

It can convert either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision) and returns an empty polygon feature for
an invalid grid reference
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
  suppressWarnings({
   library(here)
   library(fs)
  })
})
# create output directory
i_am("example.Rmd")
if (!dir_exists("output")) {dir_create("output")}

# create sf data frame and export as shape file
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(geometry = grid_reference_geometry(grid_reference)) \%>\%
  sf::st_write("output/nbn_demonstration_dataset.shp")
}
}
\seealso{
Other grid reference functions: 
\code{\link{grid_reference_easting}()},
\code{\link{grid_reference_hectad}()},
\code{\link{grid_reference_hectare}()},
\code{\link{grid_reference_monad}()},
\code{\link{grid_reference_northing}()},
\code{\link{grid_reference_pentad}()},
\code{\link{grid_reference_projection}()},
\code{\link{grid_reference_tetrad}()},
\code{\link{grid_reference_valid}()},
\code{\link{precision.gridref}()}
}
\concept{grid reference functions}
