% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{grid_reference_precision}
\alias{grid_reference_precision}
\title{Get precison for grid reference}
\usage{
grid_reference_precision(grid_reference)
}
\arguments{
\item{grid_reference}{character, British or Irish grid reference}
}
\value{
integer, precision of grid reference in metres.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function returns the grid reference's precsion in metres.
It uses the gridCoords function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.

It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add precision column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(precision = grid_reference_precision(grid_reference))
}
}
\seealso{
Other grid reference functions: 
\code{\link{grid_reference_easting}()},
\code{\link{grid_reference_geometry}()},
\code{\link{grid_reference_northing}()},
\code{\link{grid_reference_projection}()},
\code{\link{grid_reference_valid}()}
}
\concept{grid reference functions}
