% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{grid_reference_to_geometry}
\alias{grid_reference_to_geometry}
\title{Convert OSGB Grid reference to polygon geometry feature}
\usage{
grid_reference_to_geometry(grid_reference)
}
\arguments{
\item{grid_reference}{character, British or Irish grid reference}
}
\value{
geometry, square polygon feature
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function converts a grid reference to its square polygon geometry feature
through conversion to well-known text. It uses the gridCoords function in the
archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.

It can convert either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
grid_reference_to_geometry("SP123456")
}
}
