% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{grid_reference_valid}
\alias{grid_reference_valid}
\title{Check if grid reference is a valid OSGB or OSNI Grid reference}
\usage{
grid_reference_valid(grid_reference)
}
\arguments{
\item{grid_reference}{character, British or Irish grid reference}
}
\value{
logical, grid reference valid (TRUE) or invalid (FALSE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function checks if the grid reference is a valid OSGB or OSNI grid reference.
It uses the gridCoords function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.

It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add valid column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(valid = grid_reference_valid(grid_reference))
}
}
\seealso{
Other grid reference functions: 
\code{\link{grid_reference_easting}()},
\code{\link{grid_reference_geometry}()},
\code{\link{grid_reference_hectad}()},
\code{\link{grid_reference_hectare}()},
\code{\link{grid_reference_monad}()},
\code{\link{grid_reference_northing}()},
\code{\link{grid_reference_pentad}()},
\code{\link{grid_reference_projection}()},
\code{\link{grid_reference_tetrad}()},
\code{\link{precision.gridref}()}
}
\concept{grid reference functions}
