% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{gridCoords.gridref}
\alias{gridCoords.gridref}
\title{Get x,y coordinates from a grid reference}
\usage{
gridCoords.gridref(grid, units = c("km", "m"))
}
\arguments{
\item{grid}{character, an OSGB or OSNI grid reference string}

\item{units}{character, metres or kilometres (default)}
}
\value{
a list of class "gridref" with the following contents:
\itemize{
\item grid	the original grid reference
\item system	the grid reference system, either "OSGB" or "OSNI"
\item x the x coordinate (easting) in requested units
\item y the y coordinate (northing)in requested units
\item units "m" or "km"
\item precision the precision of the original grid reference in metres
}
}
\description{
Given an OSGB or OSNI grid reference string, get the x,y coordinates of the OSGB
or OSNI grid for the bottom, left-hand corner of the grid square. The units
parameter controls the units (metres m or kilometres km) in which the coordinates
should be returned

function from \href{https://github.com/ropensci-archive/rnbn/issues/37}{archived rnbn package}
}
\author{
Stuart Ball
}
