% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{gridRef}
\alias{gridRef}
\title{Generic function for gridRef}
\usage{
gridRef(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by function gridRef}
}
\value{

}
\description{
Generic function for gridRef as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
