% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{gridreference_to_wkt}
\alias{gridreference_to_wkt}
\title{Convert OSGB Gridreference to well-known text}
\usage{
gridreference_to_wkt(grid_reference)
}
\arguments{
\item{grid_reference}{character, gridreference in OSGB ()}
}
\value{
character, polygon WKT
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function converts a grid refence to well-known text, using the
gridCoords function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn}
package.

It can convert either British or Irish gridreferences up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
gridreference_to_wkt("SP123456")
}
}
