% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{gridsquare_geometry}
\alias{gridsquare_geometry}
\alias{gridsquare_geometry.gridref}
\title{Convert a OSGB or OSNI grid reference to a polygon geometry feature}
\usage{
gridsquare_geometry(x, ...)

\method{gridsquare_geometry}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string with class gridref.}
}
\value{
The square polygon geometry feature
}
\description{
This function converts a grid reference to its square polygon geometry feature
through conversion to well-known text.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision).
}
\examples{
 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset \%>\%
   clean_names() \%>\%
   slice_tail(n = 15) \%>\%
   select(grid_reference) \%>\%
   mutate(grid_reference = as_gridref(grid_reference))


# create sf data frame
suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
    library(sf)
  })
})

grid_references \%>\%
  rowwise() \%>\%
  mutate(geometry = gridsquare_geometry(grid_reference)) \%>\%
  st_as_sf()
}
\seealso{
Other grid reference functions: 
\code{\link{easting}()},
\code{\link{hectad}()},
\code{\link{hectare}()},
\code{\link{monad}()},
\code{\link{northing}()},
\code{\link{pentad}()},
\code{\link{precision}()},
\code{\link{projection}()},
\code{\link{tetrad}()}
}
\concept{grid reference functions}
