% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{hectare.gridref}
\alias{hectare.gridref}
\title{Get the 100m grid reference}
\usage{
\method{hectare}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string with class gridref.}
}
\value{
The grid reference at 100m resolution.
}
\description{
This function returns the 100m grid reference for a higher precision grid reference.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision).
}
\examples{
 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset \%>\%
   clean_names() \%>\%
   slice_tail(n = 15) \%>\%
   select(grid_reference) \%>\%
   mutate(grid_reference = as_gridref(grid_reference))


# add one_hundred_m column
suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
  })
})

grid_references \%>\%
  rowwise() \%>\%
  mutate(one_hundred_m = hectare(grid_reference))
}
\seealso{
Other grid reference functions: 
\code{\link{easting.gridref}()},
\code{\link{gridsquare_geometry.gridref}()},
\code{\link{hectad.gridref}()},
\code{\link{monad.gridref}()},
\code{\link{northing.gridref}()},
\code{\link{pentad.gridref}()},
\code{\link{precision.gridref}()},
\code{\link{projection.gridref}()},
\code{\link{tetrad.gridref}()}
}
\concept{grid reference functions}
