% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{hectare}
\alias{hectare}
\title{Generic function for hectare}
\usage{
hectare(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by generic function hectare}
}
\value{

}
\description{
Generic function for hectare as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
