% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_keyrings}
\alias{list_keyrings}
\title{List the keyrings}
\usage{
list_keyrings(keyring_name)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring.}
}
\description{
List the keyrings present in the Windows Credential Store.
}
\seealso{
Other keyring functions: 
\code{\link{add_key}()},
\code{\link{create_locked_keyring}()},
\code{\link{delete_keyring}()},
\code{\link{delete_key}()},
\code{\link{get_keyring_status}()},
\code{\link{list_keys}()},
\code{\link{lock_keyring}()},
\code{\link{unlock_keyring}()}
}
\concept{keyring functions}
