% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_keys}
\alias{list_keys}
\title{list_keys}
\usage{
list_keys(keyring_name)
}
\arguments{
\item{keyring_name}{character, keyring name}
}
\value{

}
\description{
List the current keys stored on the keyring.
Each of these keys store a database connection string.
}
\examples{
\dontrun{
list_keys(readline("keyring name: "))
}
}
