% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{lock_keyring}
\alias{lock_keyring}
\title{lock_keyring}
\usage{
lock_keyring(keyring_name)
}
\arguments{
\item{keyring_name}{character, keyring name}
}
\value{

}
\description{
Lock the keyring requiring keyring password
}
\examples{
\dontrun{
lock_keyring(readline("keyring name: "))
}
}
