% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3p.R
\name{m3p}
\alias{m3p}
\title{Continous state-trait ordinal regression}
\usage{
m3p(formula, data, id)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}

\item{data}{A data frame containing the variables in the model.}

\item{id}{A variable name for the subjects of the \code{data}.}
}
\value{
A fitted three-parameter model object with a continous response and predictor ordinal 
variables extracted from \code{formula}.
}
\description{
This function performs regression of a continuous ordinal response on an ordinal predictor compared 
with a general model in \code{\link{store}} state-trait ordinal regression.
}
\examples{
## maybe something with simulated data...
}
\seealso{
\code{\link{store}}
%% @references %% ~put references to the literature/web site here ~
}
\author{
Prince P. Osei and Philip T. Reiss
}
