% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meadian.R
\name{meadian}
\alias{meadian}
\title{Median for an ordinal variable}
\usage{
meadian(v)
}
\arguments{
\item{v}{A vector of numeric values}
}
\value{
The median as an average of the mean and default median of \code{v}.
%% @note %% ~~further notes~~
}
\description{
This function simply computes median of ordinal variable, with one modification:
if the median is between two values, the larger or smaller value is returned, depending 
if the mean is larger or smaller.
}
\examples{
data(cars)           
meadian(cars$dist)     
}
\author{
Philip T. Reiss and Prince P. Osei
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
%% @references %% ~put references to the literature/web site here ~
}
