% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{merge_pdf_documents}
\alias{merge_pdf_documents}
\title{Merge pdf documents}
\usage{
merge_pdf_documents(pdf_paths = NULL, saved_merged_pdf_path = NULL)
}
\arguments{
\item{pdf_paths}{vector of pdf document paths in the order they are to be to merged,
default NULL to select the order of the pdf document paths.}

\item{saved_merged_pdf_path}{character string of merged pdf document path, default
NULL to select merged pdf document path.}
}
\value{
vector of pdf document paths to merge
}
\description{
This function merges one or more pdf documents, saving the resulting merged
pdf document either to a provided path or default merged pdf name.
}
