% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{monad}
\alias{monad}
\title{Generic function for monad}
\usage{
monad(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by generic function monad}
}
\value{

}
\description{
Generic function for monad as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
