% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{monad_api}
\alias{monad_api}
\title{monad_api}
\usage{
monad_api(grid_reference, centre = FALSE)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string with class gridref.}
}
\value{
The grid reference at 2km resolution.
}
\description{
This function returns the 1km grid reference for a higher precision grid reference.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision).
}
\section{REST API}{

A REST API endpoint is available once a Plumber router has been created and can be
requested by adding \strong{_api} to the end of the function.

The API returns "" instead of NULL so that it may be used with the dplyr::mutate function.
Wrap the API call with dplyr::na_if to convert "" to NA.\preformatted{suppressPackageStartupMessages(\{
  library(store)
  suppressWarnings(\{
    library(plumber)
  \})
\})
pr("../R/gridref_api.R") \%>\%
  pr_run(port = 8000)
}
}

\examples{
 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset \%>\%
   clean_names() \%>\%
   slice_tail(n = 15) \%>\%
   select(grid_reference) \%>\%
   mutate(grid_reference = as_gridref(grid_reference))


# add one_km column
suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
  })
})

grid_references \%>\%
  rowwise() \%>\%
  mutate(one_km = monad(grid_reference))
}
\seealso{
Other grid reference functions: 
\code{\link{easting}()},
\code{\link{gridsquare_geometry}()},
\code{\link{hectad}()},
\code{\link{hectare}()},
\code{\link{northing}()},
\code{\link{pentad}()},
\code{\link{precision}()},
\code{\link{projection}()},
\code{\link{tetrad}()}
}
