% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{northing}
\alias{northing}
\title{Generic function for northing}
\usage{
northing(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by generic function northing}
}
\value{

}
\description{
Generic function for northing as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
