% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{pentad.gridref}
\alias{pentad.gridref}
\title{Get 5km grid reference}
\usage{
\method{pentad}{gridref}(grid)
}
\arguments{
\item{grid_reference}{character, British or Irish grid reference}
}
\value{
character, 5km grid reference
}
\description{
This function returns the 5km grid reference for a higher precision grid reference.
It uses the gridRef function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.
}
\details{
It can convert either British or Irish grid references up to 10 figure (1m precision),
but does not convert tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add five_km column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::mutate(grid_reference = as_gridref(grid_reference)) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(five_km = pentad(grid_reference))
}
}
\seealso{
Other grid reference functions: 
\code{\link{easting.gridref}()},
\code{\link{grid_reference_geometry}()},
\code{\link{hectad.gridref}()},
\code{\link{hectare.gridref}()},
\code{\link{monad.gridref}()},
\code{\link{northing.gridref}()},
\code{\link{precision.gridref}()},
\code{\link{projection.gridref}()},
\code{\link{tetrad.gridref}()}
}
\concept{grid reference functions}
