% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.store.R
\name{plot.store}
\alias{plot.store}
\title{Plotting for state-trait ordinal regression coefficients and fitted values}
\usage{
\method{plot}{store}(
  x,
  type = c("coef"),
  bubble = FALSE,
  col = NULL,
  xlab = "State",
  ylab = "Trait (person-specific median)",
  title = NULL,
  range = c(0, 25),
  hjust = 0.5,
  lineheight = 1.2,
  cex.lab = 1.5,
  cex.main = 1.5,
  ...
)
}
\arguments{
\item{x}{A model object of class \code{\link{store}}.}

\item{type}{the type of model output to plot: either \code{"coef"} (the default) to plot coefficients, or \code{"fitted"} to plot fitted values.}

\item{bubble}{Logical: a bubble plot if \code{TRUE}, otherwise an \code{\link{image}} plot.}

\item{col}{The colors for the range of the fitted values.}

\item{xlab, ylab}{x- and y-axis labels.}

\item{title}{Plot title.}

\item{range}{Numeric vector of length 2 specifying the minimum and maximum size of the bubbles, if \code{bubble==TRUE}.}

\item{hjust}{Horizontal justification (in [0,1]).}

\item{lineheight}{Line height.}

\item{cex.lab}{Magnification for x and y labels.}

\item{cex.main}{Magnification for main title.}

\item{\dots}{Other graphical parameters.}
}
\value{
None; a plot is created.
}
\description{
Plots coefficients for the various trait-state combinations of a \code{\link{store}} fit.
}
\note{
This function assumes the \code{\link{store}} object includes the underlying data 
(\code{save.data==TRUE}, the default in \code{\link{store}}).
}
\seealso{
\code{\link{store}}
}
\author{
Prince P. Osei and Philip T. Reiss
}
