% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{precision}
\alias{precision}
\title{Generic function for precision}
\usage{
precision(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by generic function precision}
}
\value{

}
\description{
Generic function for precision as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
