% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{projection.gridref}
\alias{projection.gridref}
\title{Get the projection of a grid reference}
\usage{
\method{projection}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string with class gridref.}
}
\value{
The projection of the grid reference in British National Grid (OSGB) or Irish National Grid (OSNI).
}
\description{
This function returns the grid reference's projection, either as OSGB or OSNI.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision).
}
\examples{
 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset \%>\%
   clean_names() \%>\%
   slice_tail(n = 15) \%>\%
   select(grid_reference) \%>\%
   mutate(grid_reference = as_gridref(grid_reference))


# add projection column
suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
  })
})

grid_references \%>\%
  rowwise() \%>\%
  mutate(projection = projection(grid_reference))
}
\seealso{
Other grid reference functions: 
\code{\link{easting.gridref}()},
\code{\link{gridsquare_geometry.gridref}()},
\code{\link{hectad.gridref}()},
\code{\link{hectare.gridref}()},
\code{\link{monad.gridref}()},
\code{\link{northing.gridref}()},
\code{\link{pentad.gridref}()},
\code{\link{precision.gridref}()},
\code{\link{tetrad.gridref}()}
}
\concept{grid reference functions}
