% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{projection}
\alias{projection}
\title{Generic function for projection}
\usage{
projection(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by generic function projection}
}
\value{

}
\description{
Generic function for projection as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
