% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{remove_pdf_document_pages}
\alias{remove_pdf_document_pages}
\title{Remove pdf document pages}
\usage{
remove_pdf_document_pages(pdf_path = NULL, pages_to_remove)
}
\arguments{
\item{pdf_path}{path character string of pdf document.}

\item{pages_to_remove}{numeric vector of page numbers to extract.}
}
\value{
path character string of pdf document.
}
\description{
This function removes one or more pages from a pdf document, saving the original
updated document (adding original_updated to the document name) and saving
the removed pages as a separate pdf document (adding removed_pages to the document name).
The path to the pdf document is requested if not provided.
}
\seealso{
Other pdf: 
\code{\link{count_pdf_document_pages}()},
\code{\link{merge_pdf_documents}()}
}
\concept{pdf}
