% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store}
\alias{store}
\title{State-trait ordinal regression}
\usage{
store(
  formula,
  data,
  id,
  penalize = TRUE,
  method = "mgcv",
  model = "TS",
  covt = NULL,
  save.data = TRUE,
  ...
)
}
\arguments{
\item{formula}{An expression of the form \code{y ~ x} representing the model.}

\item{data}{Data frame containing the variables in the model.}

\item{id}{Variable name for the subjects of the \code{data}.}

\item{penalize}{A logical variable: if TRUE (the default), a non-linearity penalty is applied.}

\item{method}{Package implementing ordinal mixed-effect regression: either "\code{mgcv}" (the default) or "\code{ordinal}".}

\item{model}{The model to fit: "TS" for trait-state model with random effects (the default), "trait" for trait-only model with random effects, "RE" for model with random effects only.}

\item{covt}{The covariates of the model; default is NULL.}

\item{save.data}{A logical variable: if TRUE (the default) store the \code{data}.}

\item{\ldots}{Other arguments, passed to \code{\link[mgcv]{gam}}.}
}
\value{
A fitted model object with response and predictor ordinal variables extracted from \code{formula}. 
This is a \code{\link[mgcv]{gam}} object with some additional components.

When \code{model} is \code{"TS"} or \code{"trait"}, the model matrix is called \code{tsmat}. 
Coefficient \code{tsmatmed1}, for example, is the "trait effect" associated with median predictor value 1.
Coefficient \code{tsmat1.2} is the "state effect" of current predictor value 2 for an individual whose median is 1.
}
\description{
This function performs regression of an ordinal response on an ordinal predictor with
intensive longitudinal data. The methodology allows for separating state effects from 
trait effects.
}
\examples{
\dontrun{
mod <- store(pleasant.~stressed., data=thoughts, id="Subject")  # takes several minutes
summary(mod)
plot(mod)
}
}
\author{
Prince P. Osei and Philip T. Reiss  
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
%% @references Osei, P. P. and Reiss, P. T. Ordinal state-trait regression for intensive longitudinal data. Under revision.
}
