% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{tetrad.gridref}
\alias{tetrad.gridref}
\title{Get the 2km grid reference}
\usage{
\method{tetrad}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{character gridref class, British or Irish grid reference}
}
\value{
character, 2km grid reference
}
\description{
This function returns the 2km grid reference for a higher precision grid reference.
It uses the gridRef function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.

It can convert either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add two_km column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::mutate(grid_reference = as_gridref(grid_reference)) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(two_km = tetrad(grid_reference))
}
}
\seealso{
Other grid reference functions: 
\code{\link{easting.gridref}()},
\code{\link{gridsquare_geometry.gridref}()},
\code{\link{hectad.gridref}()},
\code{\link{hectare.gridref}()},
\code{\link{monad.gridref}()},
\code{\link{northing.gridref}()},
\code{\link{pentad.gridref}()},
\code{\link{precision.gridref}()},
\code{\link{projection.gridref}()}
}
\concept{grid reference functions}
