% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{tetrad}
\alias{tetrad}
\title{Generic function for tetrad}
\usage{
tetrad(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by generic function tetrad}
}
\value{

}
\description{
Generic function for hectad as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
