% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{tidy_spatial_data}
\alias{tidy_spatial_data}
\title{Tidy spatial data}
\usage{
tidy_spatial_data(sf_data, epsg, check_valid = FALSE)
}
\arguments{
\item{sf_data}{sf object, spatial data frame to tidy}

\item{epsg}{integer, EPSG number}

\item{check_valid}{logical, whether to check for invalid geometry (TRUE) or not (default)}
}
\value{
sf object, tidied spatial data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function tidies spatial data converting the attribute columns to snake case,
set the supplied coordinate reference system, extracting the polygons from a
geometry collection and optionally checking for invalid geometry
}
\examples{
\dontrun{
tidy_spatial_data(sf_data = "uk_ireland_base_map", epsg = 27700, check_valid = TRUE)
}
}
