% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{unlock_keyring}
\alias{unlock_keyring}
\title{unlock_keyring}
\usage{
unlock_keyring(keyring_name)
}
\arguments{
\item{keyring_name}{character, keyring name}
}
\value{

}
\description{
Unlock the keyring requiring keyring password
}
\examples{
\dontrun{
unlock_keyring(readline("keyring name: "))
}
}
