% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assurance.R
\name{unnest_failed_validation_results}
\alias{unnest_failed_validation_results}
\title{Unnest failed validation results}
\usage{
unnest_failed_validation_results(report)
}
\arguments{
\item{report}{validation report created by the data.validator package}
}
\value{
tibble containing the failed validation information
}
\description{
This function unnests the validation report, created by the data.validator package,
for validation tests resulting in errors or warnings. A tibble containing the
failed information is exported.
}
