% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{validate_gridref}
\alias{validate_gridref}
\title{Validate grid reference}
\usage{
validate_gridref(grid)
}
\arguments{
\item{grid}{character, grid reference string to validate}
}
\value{
grid, validated grid reference string
}
\description{
Validator function to check that grid reference is valid. Taken from
\href{https://github.com/ropensci-archive/rnbn/issues/37}{archived rnbn package}.
Error returned if invalid
}
\author{
Stuart Ball
}
