% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{view_key}
\alias{view_key}
\title{View key details}
\usage{
view_key(keyring_name, service_name)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring.}

\item{service_name}{A character string providing the name of the key.}
}
\value{
The JSON string containing the connection string stored in the key.
}
\description{
Returns the JSON string containing the connection string stored in the key.
The password to unlock the keyring is required before the key details are returned.
}
\seealso{
Other keyring functions: 
\code{\link{add_key}()},
\code{\link{create_locked_keyring}()},
\code{\link{delete_key}()},
\code{\link{delete_keyring}()},
\code{\link{get_keyring_status}()},
\code{\link{list_keyrings}()},
\code{\link{list_keys}()},
\code{\link{lock_keyring}()},
\code{\link{unlock_keyring}()}
}
\concept{keyring functions}
