% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_key}
\alias{add_key}
\title{add_key}
\usage{
add_key(keyring_name, service_name)
}
\arguments{
\item{keyring_name, }{character keyring name}

\item{service_name, }{character key name}
}
\value{

}
\description{
Add a new key to the keyring to store a new database connection string.
If the keyring is locked then unlock it first with its password.
}
\details{
On creating the key, the name of the key relating to the database is added under the service name
and the database connections string is added as a json string in the password dialog box.

SQL Server windows authentication: {"server":"SERVER_NAME", "database":"DATABASE_NAME"}
For SQLEXPRESS on the local machine the server name is "localhost\\SQLEXPRESS"
}
\examples{
\dontrun{
add_key(readline("keyring name: "), readline("service name: "))
}
}
