% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_key}
\alias{add_key}
\title{Add a key to a keyring}
\usage{
add_key(keyring_name, service_name)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring.}

\item{service_name}{A character string providing the name of the key.}
}
\description{
Add a new key to the keyring to store a new database connection string.
If the keyring is locked then unlock it first with its password.
}
\details{
On creating the key, the name of the key relating to the database is added under the service name
and the database connections string is added as a json string in the password dialog box.
\itemize{
\item SQL Server windows authentication: \{"server":"SERVER_NAME", "database":"DATABASE_NAME"\}.
\item PostGresSQL database: \{"server":"SERVER_NAME", "database":"DATABASE_NAME", "username":"USERNAME", "password":"PASSWORD"\}.
\item For SQLEXPRESS on the local machine the server name is "localhost\\SQLEXPRESS".
}
}
\seealso{
Other keyring functions: 
\code{\link{create_locked_keyring}()},
\code{\link{delete_key}()},
\code{\link{get_keyring_status}()},
\code{\link{list_keys}()},
\code{\link{lock_keyring}()},
\code{\link{unlock_keyring}()}
}
\concept{keyring functions}
