% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{add_tibble_to_list}
\alias{add_tibble_to_list}
\title{add tibble to list}
\usage{
add_tibble_to_list(.data, .name)
}
\arguments{
\item{.data}{tibble, tibble to be added}

\item{.name}{character, name of tibble to be added}
}
\value{
tibble, original tibble is returned invisibly
}
\description{
This function prepends tibbles to a list, creating the list
in the global environment if not already present
}
\details{
The function is an example of \href{http://adv-r.had.co.nz/Environments.html}{assigning a variable to an environment}
}
\examples{
\dontrun{
add_tibble_to_list(tibble, "tibble_name")
}
}
