% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{as_gridref}
\alias{as_gridref}
\title{Convert a string to a gridref object}
\usage{
as_gridref(grid_reference = character())
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string.}
}
\value{
A grid reference character string with an added gridref class.
}
\description{
Converts a valid grid reference character string to a gridref object.
}
\details{
Checks that the character string is a valid grid reference before converting to
a gridref class. Returns an error if the grid reference is invalid.
}
\examples{
 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset \%>\%
   clean_names() \%>\%
   slice_tail(n = 15) \%>\%
   select(grid_reference) \%>\%
   mutate(grid_reference = as_gridref(grid_reference))


# check grid_reference class
suppressPackageStartupMessages({
  suppressWarnings({
    library(purrr)
  })
})

grid_references \%>\%
 map_chr(., class)
}
\seealso{
Constructor function for gridref class as described in \href{https://adv-r.hadley.nz/s3.html}{S3}
chapter of Advanced R second additon by Hadley Wickham.
}
