% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assurance.R
\name{compare_dataset_versions}
\alias{compare_dataset_versions}
\title{Compare two versions of a dataset}
\usage{
compare_dataset_versions(old_version, new_version)
}
\arguments{
\item{old_version}{The earlier version of the dataset as a data frame}

\item{new_version}{The later version of the dataset as a data frame}
}
\value{
The data frame with an additional difference column indicating new, removed or
updated rows highlighted with #
}
\description{
This function compares two versions of a dataset returning the dataset with the added, removed
or changed rows identified, using the \href{https://github.com/edwindj/daff}{daff package}.
The compared dataset can then be exported into an Excel spreadsheet to quickly
identify where values have been changed using conditional formatting, on text
containing #.
}
\details{
An initial check should be performed prior to comparing versions to check that
the column names are identical and that there has not been any addition or removal
of columns between dataset versions, so that the dataset schema can be made the same
between versions if necessary. This check can be done using the compare function
in the \href{https://www.tidyverse.org/blog/2020/10/waldo/}{waldo package}
}
\examples{
suppressPackageStartupMessages({
  library(store)
  suppressWarnings({
    library(palmerpenguins)
    library(dplyr)
  })
})
# select top 5 heaviest penguins from each species on each island
heaviest_penguins <- penguins \%>\%
  select(species, island, body_mass_g) \%>\%
  group_by(species, island) \%>\%
  arrange(desc(body_mass_g)) \%>\%
  slice_head(n = 5) \%>\%
  ungroup()
heaviest_penguins

suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
  })
})

## each version will require an unique identifier
heaviest_penguins <- heaviest_penguins \%>\%
  mutate(id = row_number()) \%>\%
  relocate(id)

## old_version: exclude Chinstrap penguins
heaviest_penguins_old <- heaviest_penguins \%>\%
  filter(species != "Chinstrap")

## new_version: exclude Gentoo penguins and convert body mass to kilograms
heaviest_penguins_new <- heaviest_penguins \%>\%
  filter(species != "Gentoo") \%>\%
  mutate(body_mass_g = body_mass_g / 1000) \%>\%
  rename(body_mass_kg = body_mass_g)

# check columns and column names are identical between versions
waldo::compare(heaviest_penguins_old, heaviest_penguins_new)

# make columns same between versions
heaviest_penguins_old <- heaviest_penguins_old \%>\%
  rename(body_mass = body_mass_g)

heaviest_penguins_new <- heaviest_penguins_new \%>\%
  rename(body_mass = body_mass_kg)

# compare versions of dataset
suppressWarnings(compare_dataset_versions(heaviest_penguins_old, heaviest_penguins_new))
}
