% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{ISO-3166 three letter country codes with United Nations regions}
\format{
A tibble with 250 rows and 3 variables
\describe{
\item{alpha_3}{ISO_3166_1 country code}
\item{country_name}{Name of country}
\item{un_region_name}{Name of United Nations region}
}
}
\usage{
country_codes
}
\description{
International Organisation for Standardisation three letter codes for the
representation of country names taken from the \href{https://cran.r-project.org/web/packages/ISOcodes/index.html}{ISOcodes package}
, associated with UN regions taken from \href{https://unstats.un.org/unsd/methodology/m49/}{United Nations website}.
}
\details{
The missing old European country Yugoslavia was added, along with the UN regions
for one Americas country BES Bonaire, Sint Eustatius and Saba, and three Asian
countries HKG Hong Kong, MAC Macao and TWN Taiwan, Province of China.
}
\keyword{datasets}
