% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{\href{tables/country_codes_table.html}{ISO-3166 three letter country codes with United Nations regions and European Union membership}}
\format{
A tibble with 250 rows and 4 variables
\describe{
\item{alpha_3}{ISO_3166_1 country code}
\item{country_name}{Name of country}
\item{un_region_name}{Name of United Nations region}
\item{eu_member}{Whether the country is a member of the European Union or not}
}
}
\usage{
country_codes
}
\description{
International Organisation for Standardisation three letter codes for the
representation of country names taken from the \href{https://cran.r-project.org/web/packages/ISOcodes/index.html}{ISOcodes package}
, associated with UN regions taken from \href{https://unstats.un.org/unsd/methodology/m49/}{United Nations website},
and whether the country is the member of the \href{https://europa.eu/european-union/about-eu/countries_en}{European Union}

The missing old European country Yugoslavia was added, along with the UN regions
for one Americas country BES Bonaire, Sint Eustatius and Saba, and three Asian
countries HKG Hong Kong, MAC Macao and TWN Taiwan, Province of China.
}
\examples{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
  library(gt)
  library(here)
  library(fs)
 })
})

# create output directory
i_am("docs/reference/country_codes.html")
if (!dir_exists("tables")) {dir_create("tables")}

# create html table
country_codes \%>\%
  gt(groupname_col = "un_region_name") \%>\%
  tab_header(title = md("**ISO-3166 three letter country codes**"),
             subtitle = md("*Includes United Nations regions and European
                                Union membership*")) \%>\%
  cols_hide(eu_member) \%>\%
  tab_style(
    style = list(
      cell_fill(color = "#000099"), # Pantone Reflex Blue
      cell_text(color = "#FFCC00")), # Pantone Yellow
    locations = cells_body(rows = eu_member == TRUE)
  ) \%>\%
 tab_options(
    heading.align = "left",
    heading.background.color = "#2C3E50",
    column_labels.hidden = TRUE,
    row_group.font.weight = "bold",
    row_group.background.color = "#2C3E50",
    table.width = pct(50),
    table.font.size = "smaller",
    data_row.padding = px(0)
  ) \%>\%
  opt_table_lines("none") \%>\%
  gtsave("tables/country_codes_table.html")
}
\keyword{datasets}
