% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_locked_keyring}
\alias{create_locked_keyring}
\title{Create a new keyring}
\usage{
create_locked_keyring(keyring_name)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring}
}
\description{
Add a password protected keyring to the Windows Credential Store.
This keyring will be used to store database connection strings.
}
\details{
On creating the keyring you will be asked for a password.
Remember this password as it will be needed everytime you access
the database connection strings stored in this keyring.
}
\seealso{
More information on securing credentials in R can be found in RStudio's
\href{https://db.rstudio.com/best-practices/managing-credentials/}{Securing Credentials}
webpage.

Other keyring functions: 
\code{\link{add_key}()},
\code{\link{delete_key}()},
\code{\link{get_keyring_status}()},
\code{\link{list_keys}()},
\code{\link{lock_keyring}()},
\code{\link{unlock_keyring}()}
}
\concept{keyring functions}
