% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{display_variable_correlation}
\alias{display_variable_correlation}
\title{display_variable_correlation}
\usage{
display_variable_correlation(.dataset)
}
\arguments{
\item{.dataset}{data frame, data frame used to display the variable statistics}
}
\value{
a corrplot object of the correlation matrix plot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Displays a plot of the correlation matrix for each variable, showing the
correlation values between each variable combination
}
\seealso{
Other exploratory data analysis: 
\code{\link{display_variable_distribution}()},
\code{\link{display_variable_outliers}()},
\code{\link{display_variable_summary}()}
}
\concept{exploratory data analysis}
