% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_data_analysis.R
\name{display_variable_summary}
\alias{display_variable_summary}
\title{display_variable_summary}
\usage{
display_variable_summary(.dataset)
}
\arguments{
\item{.dataset}{data frame, data frame to summarise}
}
\value{
returns dataset invisibly
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Exports the basic statistics of the variables within a data frame into an output
directory. This includes a plots of the prevalence of missing values and frequency
of category levels and a table image containing a summary count of the number of rows,
missing values, unique values and zero values for each variable
}
\seealso{
Other exploratory data analysis: 
\code{\link{display_variable_correlation}()},
\code{\link{display_variable_distribution}()},
\code{\link{display_variable_outliers}()}
}
\concept{exploratory data analysis}
