% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{easting}
\alias{easting}
\alias{easting.gridref}
\title{Get the easting of a grid reference}
\usage{
easting(x, ...)

\method{easting}{gridref}(grid_reference, centre = FALSE)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string with class gridref.}

\item{centre}{Should the easting be for the lower left hand corner (default FALSE) or the centre point (TRUE).}
}
\value{
The easting of grid reference in metres.
}
\description{
This function returns the grid reference's easting in metres.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision).
}
\examples{
 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset \%>\%
   clean_names() \%>\%
   slice_tail(n = 15) \%>\%
   select(grid_reference) \%>\%
   mutate(grid_reference = as_gridref(grid_reference))


# add easting column for the centre point
suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
  })
})

grid_references \%>\%
  rowwise() \%>\%
  mutate(easting = easting(grid_reference, centre = TRUE))
}
\seealso{
Other grid reference functions: 
\code{\link{gridsquare_geometry}()},
\code{\link{hectad}()},
\code{\link{hectare}()},
\code{\link{monad}()},
\code{\link{northing}()},
\code{\link{pentad}()},
\code{\link{precision}()},
\code{\link{projection}()},
\code{\link{tetrad}()}
}
\concept{grid reference functions}
