% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_mysql_connection}
\alias{get_mysql_connection}
\title{get_mysql_connection}
\usage{
get_mysql_connection(keyring_name, service_name)
}
\arguments{
\item{keyring_name}{character, keyring name}

\item{service_name}{character, name of service storing credentials}
}
\value{
connection object, MySQL connection object
}
\description{
This function initiates a MySQL connection using stored database credentials
through the keyring package. The MySQL tables are added to the connection pane.

More information on connecting to databases using R can be found in RStudios
\href{https://db.rstudio.com/}{Databases using R} webpages.
}
\examples{
\dontrun{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
   library(DBI)
   library(odbc)
 })
})
# connect to database, using keyring stored credentials
con <- get_mysql_connection(keyring_name = "keyring_name",
                            service_name = "database_name")
# list tables
dbListTables(con)

# get query
database_query <- dbGetQuery(con, "sql_statement")

# disconnect from database
dbDisconnect(con)
}
}
