% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_postgres_connection}
\alias{get_postgres_connection}
\title{Connect to a PostgreSQL database}
\usage{
get_postgres_connection(keyring_name, service_name)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring.}

\item{service_name}{A character string providing the name of the key.}
}
\value{
A PostgreSQL database connection object.
}
\description{
This function initiates a PostgreSQL connection using stored database credentials
through the keyring package.
}
\section{Connect}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{library(DBI)
library(odbc)

# connect to database, using keyring stored credentials
con <- get_postgres_connection(keyring_name = "keyring_name",
                               service_name = "database_name")
}\if{html}{\out{</div>}}
}

\section{Query}{

More information on working with databases using R can be found in RStudio's
\href{https://db.rstudio.com/}{Databases using R} webpages.

\if{html}{\out{<div class="sourceCode">}}\preformatted{# list tables
dbListTables(con)

# get table
database_table <- tbl(con, "table_name") \%>\% collect()

# get query
database_query <- dbGetQuery(con, "sql_statement")
database_query <- tbl(con, sql("sql_statement")) \%>\% collect()
}\if{html}{\out{</div>}}
}

\section{Disconnect}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{dbDisconnect(con)
}\if{html}{\out{</div>}}
}

\seealso{
Other database functions: 
\code{\link{get_mysql_connection}()},
\code{\link{get_sqlserver_connection}()}
}
\concept{database functions}
