% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_sqlserver_connection}
\alias{get_sqlserver_connection}
\title{get_sqlserver_connection}
\usage{
get_sqlserver_connection(keyring_name, service_name, keyring = TRUE)
}
\arguments{
\item{keyring_name}{character, keyring name}

\item{service_name}{character, name of service storing credentials}

\item{keyring}{logical, use of keyring TRUE (default) or FALSE}
}
\value{
connection object, SQL Server connection object
}
\description{
This function initiates a SQL Server connection either
using stored database credentials using the keyring package
or else requesting database credentials. The SQKL Server tables
are added to the connection pane
}
\examples{
\dontrun{
get_sqlserver_connection("keyring", "database_service_name")
}
}
