% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_sqlserver_connection}
\alias{get_sqlserver_connection}
\title{Connect to a SQL Server database}
\usage{
get_sqlserver_connection(keyring_name, service_name, keyring = TRUE)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring.}

\item{service_name}{A character string providing the name of the key.}

\item{keyring}{default TRUE uses the keyring, FALSE requests credentials.}
}
\value{
A SQL Server database connection object.
}
\description{
This function initiates a SQL Server connection using windows authentication,
either using stored database credentials through the keyring package or else
requesting database credentials in the console. The SQL Server tables are added
to the connection pane.
}
\section{Connect}{
\preformatted{library(DBI)
library(odbc)

# connect to database, using keyring stored credentials
con <- get_sqlserver_connection(keyring_name = "keyring_name",
                                service_name = "database_name",
                                keyring = TRUE)
}
}

\section{Query}{

More information on working with databases using R can be found in RStudio's
\href{https://db.rstudio.com/}{Databases using R} webpages.\preformatted{# list tables
dbListTables(con)

# get table
database_table <- tbl(con, "table_name") \%>\% collect()

# get query
database_query <- dbGetQuery(con, "sql_statement")
database_query <- tbl(con, sql("sql_statement")) \%>\% collect()
}
}

\section{Disconnect}{
\preformatted{dbDisconnect(con)
}
}

\seealso{
Other database functions: 
\code{\link{get_mysql_connection}()},
\code{\link{get_postgres_connection}()}
}
\concept{database functions}
