% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_sqlserver_connection}
\alias{get_sqlserver_connection}
\title{get_sqlserver_connection}
\usage{
get_sqlserver_connection(keyring_name, service_name, keyring = TRUE)
}
\arguments{
\item{keyring_name}{character, keyring name}

\item{service_name}{character, name of service storing credentials}

\item{keyring}{logical, use of keyring TRUE (default) or request credentials (FALSE)}
}
\value{
connection object, SQL Server connection object
}
\description{
This function initiates a SQL Server connection using windows authentication,
either using stored database credentials through the keyring package or else
requesting database credentials in the console. The SQL Server tables are added
to the connection pane.
}
\details{
More information on connecting to databases using R can be found in RStudios
\href{https://db.rstudio.com/}{Databases using R} webpages.
}
\examples{
\dontrun{
suppressPackageStartupMessages({
 library(store)
 suppressWarnings({
   library(DBI)
   library(odbc)
 })
})
# connect to database, using keyring stored credentials
con <- get_sqlserver_connection(keyring_name = "keyring_name",
                                service_name = "database_name",
                                keyring = TRUE)
# list tables
dbListTables(con)

# get query
database_query <- dbGetQuery(con, "sql_statement")

# disconnect from database
dbDisconnect(con)
}
}
