% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{gridCoords}
\alias{gridCoords}
\title{Generic function for gridCoords}
\usage{
gridCoords(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by function gridCoords}
}
\value{

}
\description{
Generic function for gridCoords as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
