% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{gridCoords}
\alias{gridCoords}
\title{Get the x, y coordinates from a grid reference}
\usage{
gridCoords(x, ...)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string with class gridref.}

\item{units}{Should the returned coordinates be in metres (m) or kilometres (km).}
}
\value{
A list of class gridref with the following contents:
\itemize{
\item grid_reference: the original grid reference.
\item system:	the grid reference projection system, either "OSGB" or "OSNI".
\item x: the easting coordinate in requested units.
\item y: the northing coordinate in requested units.
\item units: the requested units either metres (m) or kilometres (km).
\item precision: the precision of the original grid reference in metres
}
}
\description{
Given an OSGB or OSNI grid reference string, get the x, y coordinates of the OSGB
or OSNI grid for the bottom, left-hand corner of the grid square. The units
parameter controls the units (metres m or kilometres km) in which the coordinates
should be returned.
}
\seealso{
Function taken from \href{https://github.com/ropensci-archive/rnbn/issues/37}{archived rnbn package}.
}
\author{
Stuart Ball.
}
\keyword{internal}
