% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{gridRef}
\alias{gridRef}
\title{Manipulate OSGB or OSNI grid reference string}
\usage{
gridRef(format)
}
\arguments{
\item{format}{character, the format you want back. The possibilities are: sq10km,
sq5km, tetrad, sq1km, sq100m, sq10m}

\item{grid}{character, the grid reference to be manipulated}
}
\value{
a list of class "gridref" with the following contents:
\itemize{
\item the original grid reference
\item the grid reference string formatted as requested
\item the grid reference system, either "OSGB" or "OSNI"
\item the precision of the formatted grid reference in metres
}
}
\description{
Extracts grid reference strings at various precisions from the supplied grid
reference string - if possible! For example, if you supply a 1km square reference
TL2998, then you could get the 10km square TL29, but not a 100m square grid reference.

function from \href{https://github.com/ropensci-archive/rnbn/issues/37}{archived rnbn package}
}
\seealso{
Tetrads are 2x2km squares and are often used for mapping distributions at a the
scale of a county or similar sized local area. They are labelled using the 10km
square followed by a single, upper-case letter (since there are 25 tetrads in a
10km square, the letter "O" is not used to avoid confusion with zero). This is
named the DINTY system after the letters in the second row of this table.

5x5km squares (sometimes called "pentads") are used for mapping at a regional scale.
They are labelled using the name of the 10km square followed by two upper-case letters.
}
\author{
Stuart Ball
}
