% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{gridsquare_geometry}
\alias{gridsquare_geometry}
\title{Generic function for gridsquare_geometry}
\usage{
gridsquare_geometry(x, ...)
}
\arguments{
\item{x}{object, contains class for method dispatch by generic function gridsquare_geometry}
}
\value{

}
\description{
Generic function for gridsquare_geometry as described in \href{http://adv-r.had.co.nz/S3.html}{The S3 Object system}
chapter of Advanced R by Hadley Wickham
}
