% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{list_keys}
\alias{list_keys}
\title{List the keys within a keyring}
\usage{
list_keys(keyring_name)
}
\arguments{
\item{keyring_name}{A character string providing the name of the keyring.}
}
\value{
A data frame listing the keys in the keyring.
}
\description{
List the current keys stored on the keyring.
Each of these keys store a database connection string.
}
\seealso{
Other keyring functions: 
\code{\link{add_key}()},
\code{\link{create_locked_keyring}()},
\code{\link{delete_keyring}()},
\code{\link{delete_key}()},
\code{\link{get_keyring_status}()},
\code{\link{list_keyrings}()},
\code{\link{lock_keyring}()},
\code{\link{unlock_keyring}()},
\code{\link{view_key}()}
}
\concept{keyring functions}
