% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{lock_keyring}
\alias{lock_keyring}
\title{Lock a keyring}
\usage{
lock_keyring(keyring_name)
}
\arguments{
\item{keyring_name}{character, keyring name}
}
\value{

}
\description{
Lock the keyring requiring keyring password
}
\examples{
\dontrun{
lock_keyring(readline("keyring name: "))
}
}
\seealso{
Other keyring functions: 
\code{\link{add_key}()},
\code{\link{create_locked_keyring}()},
\code{\link{delete_key}()},
\code{\link{get_keyring_status}()},
\code{\link{list_keys}()},
\code{\link{unlock_keyring}()}
}
\concept{keyring functions}
