% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3p.R
\name{m3p}
\alias{m3p}
\title{Three-parameter state-trait ordinal regression}
\usage{
m3p(formula, data, id)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}

\item{data}{A data frame containing the variables in the model.}

\item{id}{A variable name for the subjects of the \code{data}.}
}
\value{
A fitted three-parameter model object with a continous response and predictor ordinal 
variables extracted from \code{formula}.
}
\description{
This function fits a "vanilla" regression model for a repeated ordinal response and an
ordinal predictor. On the latent scale this is a linear model with a random person effect
and three fixed effects:
the intercept, the effect of the trait (person-level median of the predictor), 
and the effect of the deviation of state (current predictor value) from trait. This model
model serves as a baseline with which the more flexible model implemented by \code{\link{store}} may be compared.
See Osei and Reiss (2022) for details.
}
\examples{
## maybe something with simulated data...
}
\references{
Osei, P. P. and Reiss, P. T. (2022). Ordinal state-trait regression for intensive longitudinal data. Under revision.
}
\seealso{
\code{\link{store}}
}
\author{
Prince P. Osei and Philip T. Reiss
}
