% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{pentad}
\alias{pentad}
\alias{pentad.gridref}
\title{Get the 5km grid reference}
\usage{
pentad(x, ...)

\method{pentad}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string with class gridref.}
}
\value{
The grid reference at 5km resolution.
}
\description{
This function returns the 5km grid reference for a higher precision grid reference.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision).
}
\examples{
 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset \%>\%
   clean_names() \%>\%
   slice_tail(n = 15) \%>\%
   select(grid_reference) \%>\%
   mutate(grid_reference = as_gridref(grid_reference))


# add five_km column
suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
  })
})

grid_references \%>\%
  rowwise() \%>\%
  mutate(five_km = pentad(grid_reference))
}
\seealso{
Other grid reference functions: 
\code{\link{easting}()},
\code{\link{gridsquare_geometry}()},
\code{\link{hectad}()},
\code{\link{hectare}()},
\code{\link{monad}()},
\code{\link{northing}()},
\code{\link{precision}()},
\code{\link{projection}()},
\code{\link{tetrad}()}
}
\concept{grid reference functions}
