% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{precision.gridref}
\alias{precision.gridref}
\title{Get precision for grid reference}
\usage{
\method{precision}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{character, British or Irish grid reference}
}
\value{
integer, precision of grid reference in metres.
}
\description{
This function returns the grid reference's precision in metres.
It uses the gridCoords function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.

It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add precision column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::mutate(grid_reference = as_gridref(grid_reference)) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(precision = precision(grid_reference))
}
}
\seealso{
Other grid reference functions: 
\code{\link{easting.gridref}()},
\code{\link{grid_reference_geometry}()},
\code{\link{hectad.gridref}()},
\code{\link{hectare.gridref}()},
\code{\link{monad.gridref}()},
\code{\link{northing.gridref}()},
\code{\link{pentad.gridref}()},
\code{\link{projection.gridref}()},
\code{\link{tetrad.gridref}()}
}
\concept{grid reference functions}
