% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{projection.gridref}
\alias{projection.gridref}
\title{Get projection for grid reference}
\usage{
\method{projection}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{character, British or Irish grid reference}
}
\value{
character, grid reference projection in British National Grid (OSGB) or Irish National Grid (OSNI).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function returns the grid reference's projection, either as OSGB or OSNI.
It uses the gridCoords function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.

It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add projection column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::mutate(grid_reference = as_gridref(grid_reference)) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(projection = projection(grid_reference))
}
}
\seealso{
Other grid reference functions: 
\code{\link{grid_reference_easting}()},
\code{\link{grid_reference_geometry}()},
\code{\link{grid_reference_hectad}()},
\code{\link{grid_reference_hectare}()},
\code{\link{grid_reference_monad}()},
\code{\link{grid_reference_northing}()},
\code{\link{grid_reference_pentad}()},
\code{\link{grid_reference_tetrad}()},
\code{\link{precision.gridref}()}
}
\concept{grid reference functions}
