% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{projection.gridref}
\alias{projection.gridref}
\title{Get projection for grid reference}
\usage{
\method{projection}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{character gridref class, British or Irish grid reference}
}
\value{
character, grid reference projection in British National Grid (OSGB) or Irish National Grid (OSNI).
}
\description{
This function returns the grid reference's projection, either as OSGB or OSNI.
It uses the gridCoords function in the archived \href{https://github.com/ropensci-archive/rnbn/issues/37}{rnbn} package.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision)
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
})

# add projection column
nbn_demonstration_dataset \%>\%
  janitor::clean_names() \%>\%
  dplyr::select(grid_reference) \%>\%
  dplyr::mutate(grid_reference = as_gridref(grid_reference)) \%>\%
  dplyr::rowwise() \%>\%
  dplyr::mutate(projection = projection(grid_reference))
}
}
\seealso{
Other grid reference functions: 
\code{\link{easting.gridref}()},
\code{\link{gridsquare_geometry.gridref}()},
\code{\link{hectad.gridref}()},
\code{\link{hectare.gridref}()},
\code{\link{monad.gridref}()},
\code{\link{northing.gridref}()},
\code{\link{pentad.gridref}()},
\code{\link{precision.gridref}()},
\code{\link{tetrad.gridref}()}
}
\concept{grid reference functions}
