% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{reduce_image_size}
\alias{reduce_image_size}
\title{reduce_image_size}
\usage{
reduce_image_size(path_image, image_size = "50\%")
}
\arguments{
\item{path_image}{character, path name of image}

\item{image_size}{percentage size reduction, default = 50\%}
}
\value{
character, path name of image returned invisibly
}
\description{
This function reduces the size of an image, archiving the
original image in a separate archive sub-directory

The example batch runs the function and is an example of parallelization using
the \href{https://github.com/HenrikBengtsson/future}{future} and \href{https://davisvaughan.github.io/furrr/}{furrr}
packages
}
\examples{
\dontrun{
suppressPackageStartupMessages({
  library(store)
  suppressWarnings({
    library(fs)
    library(here)
    library(future)
    library(furrr)
  })
})

i_am("display.R")
plan(multiprocess)
here("figures") \%>\%
  dir_ls(., glob = "*.png") \%>\%
  future_map(reduce_image_size,
             image_size = "25\%",
             .options = furrr_options(seed = TRUE),
             .progress = TRUE)

}
}
