% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store}
\alias{store}
\title{State-trait ordinal regression}
\usage{
store(formula, data, id, pen = "none", method = "mgcv", covt = NULL)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}}.}

\item{data}{A data frame containing the variables in the model.}

\item{id}{A variable name for the subjects of the \code{data}.}

\item{pen}{The method of penalization to be used. If it is "none", no penalization.}

\item{method}{Package implementing ordinal mixed-effect regression: either "mgcv" (the default) or "ordinal".}

\item{covt}{The covariance structure of the model; default is NULL.}
}
\value{
A fitted model object with response and predictor ordinal variables extracted from \code{formula}.
%% @note %% ~~further notes~~
}
\description{
This function performs regression of an ordinal response on an ordinal predictor with
intensive longitudinal data. The methodology allows for separating state effects from 
trait effects.
}
\examples{
## maybe something with simulated data...
}
\author{
Philip T. Reiss and Prince P. Osei
%% @seealso %% ~~objects to See Also as \code{\link{help}}, ~~~
%% @references %% ~put references to the literature/web site here ~
}
