% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{tetrad}
\alias{tetrad}
\alias{tetrad.gridref}
\title{Get the 2km grid reference}
\usage{
tetrad(x, ...)

\method{tetrad}{gridref}(grid_reference)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string with class gridref.}
}
\value{
The grid reference at 2km resolution.
}
\description{
This function returns the 2km grid reference for a higher precision grid reference.
}
\details{
It can check either British or Irish grid references up to 10 figure (1m precision),
including tetrads (2000m precision).
}
\examples{
 suppressPackageStartupMessages({
   library(store)
   suppressWarnings({
     library(dplyr)
     library(janitor)
   })
 })

 grid_references <- nbn_demonstration_dataset \%>\%
   clean_names() \%>\%
   slice_tail(n = 15) \%>\%
   select(grid_reference) \%>\%
   mutate(grid_reference = as_gridref(grid_reference))


# add two_km column
suppressPackageStartupMessages({
  suppressWarnings({
    library(dplyr)
  })
})

grid_references \%>\%
  rowwise() \%>\%
  mutate(two_km = tetrad(grid_reference))
}
\seealso{
Other grid reference functions: 
\code{\link{easting}()},
\code{\link{gridsquare_geometry}()},
\code{\link{hectad}()},
\code{\link{hectare}()},
\code{\link{monad}()},
\code{\link{northing}()},
\code{\link{pentad}()},
\code{\link{precision}()},
\code{\link{projection}()}
}
\concept{grid reference functions}
