% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridref.R
\name{validate_gridref}
\alias{validate_gridref}
\title{Validate a grid reference}
\usage{
validate_gridref(grid_reference)
}
\arguments{
\item{grid_reference}{A Great British or Irish grid reference character string}
}
\value{
A validated grid reference character string
}
\description{
Internal validator function to check that a grid reference is valid before
converting to a gridref class. Returns an error if the grid reference is invalid
}
\details{
Taken from \href{https://github.com/ropensci-archive/rnbn/issues/37}{archived rnbn package}.
}
\author{
Stuart Ball
}
